requirements-aws.txt
# AWS AgentCore additional dependencies
boto3>=1.34.0
botocore>=1.34.0
aws-xray-sdk>=2.12.0
opentelemetry-api>=1.24.0
opentelemetry-sdk>=1.24.0
opentelemetry-exporter-otlp>=1.24.0
opentelemetry-instrumentation-fastapi>=0.45b0
opentelemetry-instrumentation-httpx>=0.45b0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
structlog>=24.1.0
prometheus-client>=0.20.0
